Some good updates
Updated version.
Notable: input images should be in the .png format to avoid some trouble with video rendering. The outputed movie won't be viewable with Windows Media Player. Oh no.
Anyway, the program now works good enough, however:
it expects there to be less than 1000 files. This seems easily fixable.
It rewrites new files every time the program is ran again. It should be able to recognize it had already done that and skip this step.